move tests to correct folder; change test and use sentinel value
