+++ tests/conversion/management/commands/result_harvester_test.py
@@ -1,4 +1,53 @@
+from collections import namedtuple
 from unittest.mock import Mock, MagicMock, patch
+
+import pytest
+
+from osmaxx.conversion_api.statuses import STARTED
+
+
+@pytest.fixture
+def queue(fake_rq_id):
+    Queue = namedtuple('Queue', ['job_ids', 'fetch_job'])
+    Job = namedtuple('Job', ['status'])
+
+    def fetch_job(job_id):
+        job = Job(status=STARTED)
+        return job
+    queue = Queue(job_ids=[str(fake_rq_id)], fetch_job=fetch_job)
+    return queue
+
+
+def test_handle_failed_jobs_calls_update_job(mocker, fake_rq_id, queue):
+    from osmaxx.conversion.management.commands import result_harvester
+    mocker.patch('django_rq.get_failed_queue', return_value=queue)
+    cmd = result_harvester.Command()
+    _update_job_mock = mocker.patch.object(cmd, '_update_job')
+    cmd._handle_failed_jobs()
+    assert _update_job_mock.call_count == 1
+    _update_job_mock.assert_called_once_with(job_id=str(fake_rq_id), queue=queue)
+
+
+@pytest.mark.django_db()
+def test_handle_successfull_jobs_calls_update_job(mocker, queue, started_conversion_job):
+    from osmaxx.conversion.management.commands import result_harvester
+    mocker.patch('django_rq.get_queue', return_value=queue)
+    cmd = result_harvester.Command()
+    _update_job_mock = mocker.patch.object(cmd, '_update_job')
+    cmd._handle_running_jobs()
+    assert _update_job_mock.call_count == 1
+    _update_job_mock.assert_called_once_with(job_id=str(started_conversion_job.rq_job_id), queue=queue)
+
+
+@pytest.mark.django_db()
+def test_handle_update_job_informs(mocker, queue, fake_rq_id, started_conversion_job):
+    from osmaxx.conversion.management.commands import result_harvester
+    mocker.patch('django_rq.get_queue', return_value=queue)
+    cmd = result_harvester.Command()
+    _update_job_mock = mocker.patch.object(cmd, '_notify')
+    cmd._update_job(job_id=fake_rq_id, queue=queue)
+    assert _update_job_mock.call_count == 1
+    _update_job_mock.assert_called_once_with(started_conversion_job)
 
 
 def test_add_meta_data_to_job_with_out_of_bounds_exception():
@@ -8,10 +57,11 @@ def test_add_meta_data_to_job_with_out_of_bounds_exception():
     conversion_job.extraction_duration = None
     conversion_job.estimated_pbf_size = None
     conversion_job.parametrization.clipping_area.clipping_multi_polygon.extent = -70, 70, 80, -10
+    sentinel_value = 42
     rq_job = Mock()
     rq_job.meta = {
-        'unzipped_result_size': 0,
-        'duration': 0,
+        'unzipped_result_size': sentinel_value,
+        'duration': sentinel_value,
     }
 
     with patch('osmaxx.conversion.management.commands.result_harvester.logger') as logger_mock:
@@ -20,6 +70,6 @@ def test_add_meta_data_to_job_with_out_of_bounds_exception():
             rq_job=rq_job
         )
         logger_mock.exception.assert_called_once_with("pbf estimation failed")
-        assert conversion_job.unzipped_result_size == 0
-        assert conversion_job.extraction_duration == 0
+        assert conversion_job.unzipped_result_size == sentinel_value
+        assert conversion_job.extraction_duration == sentinel_value
         assert conversion_job.estimated_pbf_size is None
